+++ src/adapter/console/exceptionMessage.ts
@@ -4,6 +4,7 @@
 
 import Cdp from '../../cdp/api';
 import { once } from '../../common/objUtils';
+import { StackTraceParser } from '../../common/stackTraceParser';
 import Dap from '../../dap/api';
 import { previewException } from '../objectPreview';
 import { StackTrace } from '../stackTrace';
@@ -48,11 +49,19 @@ export class ExceptionMessage extends TextualMessage<Cdp.Runtime.ExceptionDetail
 
     let message = preview.title;
     if (!message.startsWith('Uncaught')) {
-      message = 'Uncaught ' + message;
+      message = `Uncaught ${this.event.exception?.className ?? 'Error'} ` + message;
     }
 
     const stackTrace = this.stackTrace(thread);
     const args = this.event.exception && !preview.stackTrace ? [this.event.exception] : [];
+
+    // If there is a stacktrace in the exception message, beautiful its paths.
+    // If there isn't (and there isn't always) then add one.
+    if (StackTraceParser.isStackLike(message)) {
+      message = await thread.replacePathsInStackTrace(message);
+    } else if (stackTrace) {
+      message += '\n' + (await stackTrace.formatAsNative());
+    }
 
     return {
       category: 'stderr',